Fix test imports and re-organize
