+++ packages/form-generation-tool/src/generators/Bootstrap4FormGenerator.ts
@@ -21,8 +21,8 @@ import unescape from "lodash/unescape";
 import { FormAssetType, FormAsset, FormStyle, FormConfiguration, FormGenerator, FormSchema } from "../types";
 import { renderForm } from "@kie-tools/uniforms-bootstrap4-codegen/dist";
 import JSONSchemaBridge from "uniforms-bridge-json-schema";
-import { getUniformsSchema } from "./getUniformsSchema";
-import { inputSanitizationUtil } from "./InputSanitizationUtil";
+import { getUniformsSchema } from "../getUniformsSchema";
+import { inputSanitizationUtil } from "../inputSanitizationUtil";
 
 export const BOOTSTRAP4_CSS_URL = "https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css";
 export const BOOTSTRAP4_JS_URL = "https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.bundle.min.js";
