+++ packages/form-generation-tool/src/inputSanitizationUtil.ts
