--- packages/form-generation-tool/tests/generation/fs/resources/full/schemas/travels_ApplyForVisa.json
@@ -1,50 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "type": "object",
-  "properties": {
-    "traveller": {
-      "type": "object",
-      "properties": {
-        "address": {
-          "type": "object",
-          "properties": {
-            "city": { "type": "string" },
-            "country": { "type": "string" },
-            "street": { "type": "string" },
-            "zipCode": { "type": "string" }
-          }
-        },
-        "email": { "type": "string" },
-        "firstName": { "type": "string" },
-        "lastName": { "type": "string" },
-        "nationality": { "type": "string" }
-      },
-      "input": true
-    },
-    "trip": {
-      "type": "object",
-      "properties": {
-        "begin": { "type": "string", "format": "date-time" },
-        "city": { "type": "string" },
-        "country": { "type": "string" },
-        "end": { "type": "string", "format": "date-time" },
-        "visaRequired": { "type": "boolean" }
-      },
-      "input": true
-    },
-    "visaApplication": {
-      "type": "object",
-      "properties": {
-        "approved": { "type": "boolean" },
-        "city": { "type": "string" },
-        "country": { "type": "string" },
-        "duration": { "type": "integer" },
-        "firstName": { "type": "string" },
-        "lastName": { "type": "string" },
-        "nationality": { "type": "string" },
-        "passportNumber": { "type": "string" }
-      },
-      "output": true
-    }
-  }
-}
